# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Produto",
        "description": "API REST para gerenciamento de produtos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local (Docker)"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Registrar novo usuário",
                "operationId": "e911bfed995e01d2b430d00b386d43c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "User test"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Autenticar usuário",
                "operationId": "46e8c10da6f4d98e8f2702a6632b23a3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado"
                    },
                    "422": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Logout de usuário autenticado",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Logout realizado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Obter perfil do usuário autenticado",
                "operationId": "833079f8cee1115ad359b3d863d49ace",
                "responses": {
                    "200": {
                        "description": "Perfil do usuário recuperado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/produtos": {
            "get": {
                "tags": [
                    "ProdutoController"
                ],
                "summary": "Listar todos os produtos",
                "operationId": "6a5405867c96fc1f274b0fcd379441e3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produtos listados"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "404": {
                        "description": "Nenhum produto encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ProdutoController"
                ],
                "summary": "Criar um novo produto",
                "operationId": "9c3a31cb5c64ba3c23f9ce329cea097c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "valor",
                                    "quantidade",
                                    "fora_validade"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Notebook Gamer"
                                    },
                                    "valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3000
                                    },
                                    "quantidade": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "fora_validade": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/buscar-produtos": {
            "get": {
                "tags": [
                    "ProdutoController"
                ],
                "summary": "Buscar produtos por nome",
                "operationId": "ac81c8e96d73f6738b866ceb8b78805e",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados encontrados"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "404": {
                        "description": "Nenhum resultado encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/produtos/{id}": {
            "get": {
                "tags": [
                    "ProdutoController"
                ],
                "summary": "Mostrar produto por ID",
                "operationId": "835c169c4456d1a590e63c69a530aeed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ProdutoController"
                ],
                "summary": "Atualizar produto",
                "operationId": "c45ef52a7b159513ec0c8878561a6c38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Notebook Gamer"
                                    },
                                    "valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4999.9
                                    },
                                    "quantidade": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "fora_validade": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Produto atualizado"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ProdutoController"
                ],
                "summary": "Deletar produto",
                "operationId": "0ec9ccb50e373a2fd27969575616164a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto deletado"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "AuthController",
            "description": "AuthController"
        },
        {
            "name": "ProdutoController",
            "description": "ProdutoController"
        }
    ]
}